+++ NEWS.md
@@ -2,7 +2,7 @@ CohortGenerator 0.4.0
 =======================
 - Update dependency versions in DESCRIPTION to resolve issues when using older versions of readr/stringi. (Issue #13)
 - Breaking change: change the defaults for saving/getting a cohortDefinitionSet to/from the file system. (Issue #16)
-- Bug fixes
+- `getCohortCounts` now supports an optional parameter `cohortDefinitionSet` which will join the cohort counts with the cohort definition set when desired (Issue #14)
 
 CohortGenerator 0.3.0
 =======================